+++ backend/shared/models/trade.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 
 from app.core.database import Base
 from sqlalchemy import DateTime, ForeignKey, Numeric, Text, func
@@ -12,7 +13,7 @@ class Trade(Base):
     portfolio_id: Mapped[int] = mapped_column(ForeignKey('portfolios.id', ondelete='CASCADE'))
     asset_id: Mapped[int] = mapped_column(ForeignKey('assets.id'))
     direction: Mapped[str] = mapped_column(Text)  # buy/sell
-    quantity: Mapped[float] = mapped_column(Numeric, nullable=False)
-    price: Mapped[float] = mapped_column(Numeric(20, 6), nullable=False)
+    quantity: Mapped[Decimal] = mapped_column(Numeric, nullable=False)
+    price: Mapped[Decimal] = mapped_column(Numeric(20, 6), nullable=False)
     trade_time: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now())
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now())
